requirements.txt
requests==2.28.1
pyperclip==1.8.2
kivy==2.2.1
python-dotenv==1.0.0